requirements.txt
fal_client
fpdf
openai
Pillow
requests
streamlit